Move HtmlElementLookup to visu-utils.ts
